[REFACTOR] DaoConnectionImpl #7
