+++ src/main/java/com/test/controller/OtherTablesController.java
@@ -4,7 +4,6 @@ import com.test.service.UserServiceOtherTables;
 import com.test.service.UserServicePersonal;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
