+++ src/main/java/com/test/dao/DaoConnection.java
@@ -1,10 +1,13 @@
 package com.test.dao;
 
 import javax.sql.DataSource;
+import java.sql.Connection;
 
 public interface DaoConnection {
-    void connect();
-    void disconnect();
 
     DataSource getDataSource();
+
+    Connection getConnection();
+
+    void connect();
 }
