+++ src/main/java/com/test/service/UserServiceOtherTablesImpl.java
@@ -1,6 +1,6 @@
 package com.test.service;
 
-import com.test.dao.DaoFind;
+import com.test.dao.DaoOtherTables;
 import com.test.model.*;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,40 +12,40 @@ import java.util.List;
 public class UserServiceOtherTablesImpl implements UserServiceOtherTables {
     private static Logger logger = Logger.getLogger(UserServiceOtherTablesImpl.class);
 
-    private DaoFind daoFind;
+    private DaoOtherTables daoOtherTables;
 
     @Autowired
-    public void setDaoFind(DaoFind daoFind) {
-        this.daoFind = daoFind;
+    public void setDaoOtherTables(DaoOtherTables daoOtherTables) {
+        this.daoOtherTables = daoOtherTables;
     }
 
     @Override
     public List<Department> getAllDepartments() {
         logger.debug("called getAllDepartments() method. UserServiceImpl.class");
-        return daoFind.getAllDepartments();
+        return daoOtherTables.getAllDepartments();
     }
 
     @Override
     public List<Diagnosis> getAllDiagnosis() {
         logger.debug("called getAllDiagnosis() method. UserServiceImpl.class");
-        return daoFind.getAllDiagnosis();
+        return daoOtherTables.getAllDiagnosis();
     }
 
     @Override
     public List<Jobs> getAllJobs() {
         logger.debug("called getAllJobs() method. UserServiceImpl.class");
-        return daoFind.getAllJobs();
+        return daoOtherTables.getAllJobs();
     }
 
     @Override
     public List<Medicine> getAllMedicines() {
         logger.debug("called getAllMedicines() method. UserServiceImpl.class");
-        return daoFind.getAllMedicines();
+        return daoOtherTables.getAllMedicines();
     }
 
     @Override
     public List<Treatment> getAllTreatments() {
         logger.debug("called getAllTreatments() method. UserServiceImpl.class");
-        return daoFind.getAllTreatments();
+        return daoOtherTables.getAllTreatments();
     }
 }
