+++ src/main/java/com/test/service/UserServicePatientImpl.java
@@ -1,7 +1,6 @@
 package com.test.service;
 
-import com.test.dao.DaoChange;
-import com.test.dao.DaoFind;
+import com.test.dao.DaoPatients;
 import com.test.model.Patient;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,60 +13,54 @@ public class UserServicePatientImpl implements UserServicePatient {
 
     private static Logger logger = Logger.getLogger(UserServicePatientImpl.class);
 
-    private DaoFind daoFind;
+    private DaoPatients daoPatients;
 
     @Autowired
-    public void setDaoFind(DaoFind daoFind) {
-        this.daoFind = daoFind;
+    public void setDaoPatients(DaoPatients daoPatients) {
+        this.daoPatients = daoPatients;
     }
 
-    private DaoChange daoChange;
-
-    @Autowired
-    public void setDaoChange(DaoChange daoChange) {
-        this.daoChange = daoChange;
-    }
 
     @Override
     public List<Patient> allPatient() {
         logger.debug("called allPatient() method. UserServiceImpl.class");
-        return daoFind.selectAllPatient();
+        return daoPatients.selectAllPatient();
     }
 
     @Override
     public void updatePatient(int id, int diagnosisId, int medicineId) {
         logger.debug("called updatePatient() method. UserServiceImpl.class");
-        daoChange.updatePatient(id, diagnosisId, medicineId);
+        daoPatients.updatePatient(id, diagnosisId, medicineId);
     }
 
     @Override
     public Patient getByIdPatient(int id) {
         logger.debug("called getByIdPatient() method. UserServiceImpl.class");
-        return daoFind.findByIdPatient(id);
+        return daoPatients.findByIdPatient(id);
     }
 
     @Override
     public void deletePatient(int id) {
         logger.debug("called deletePatient() method. UserServiceImpl.class");
-        daoChange.deletePatient(id);
+        daoPatients.deletePatient(id);
     }
 
     @Override
     public List<Patient> getByIdPatientList(int id) {
         logger.debug("called getByIdPatientList() method. UserServiceImpl.class");
-        return daoFind.getByIdPatientList(id);
+        return daoPatients.getByIdPatientList(id);
     }
 
     @Override
     public List<Patient> findByLastNamePatient(String lastName) {
         logger.debug("called findByLastNamePatient() method. UserServiceImpl.class");
-        return daoFind.findByLastNamePatient(lastName);
+        return daoPatients.findByLastNamePatient(lastName);
     }
 
     @Override
     public void createPatient(String firstName, String lastName, String position, int phone,
                               String address, int diagnosisId, int medicineId) {
         logger.debug("called createPatient() method. UserServiceImpl.class");
-        daoChange.createPatient(firstName, lastName, position, phone, address, diagnosisId, medicineId);
+        daoPatients.createPatient(firstName, lastName, position, phone, address, diagnosisId, medicineId);
     }
 }
