+++ src/main/java/com/test/service/UserServicePersonalImpl.java
@@ -1,8 +1,7 @@
 package com.test.service;
 
-import com.test.dao.DaoChange;
-import com.test.dao.DaoFind;
-import com.test.model.*;
+import com.test.dao.DaoPersonal;
+import com.test.model.Personal;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,24 +13,17 @@ public class UserServicePersonalImpl implements UserServicePersonal {
     private static Logger logger = Logger.getLogger(UserServicePersonalImpl.class);
 
 
-    private DaoFind daoFind;
+    private DaoPersonal daoPersonal;
 
     @Autowired
-    public void setDaoFind(DaoFind daoFind) {
-        this.daoFind = daoFind;
-    }
-
-    private DaoChange daoChange;
-
-    @Autowired
-    public void setDaoChange(DaoChange daoChange) {
-        this.daoChange = daoChange;
+    public void setDaoPersonal(DaoPersonal daoPersonal) {
+        this.daoPersonal = daoPersonal;
     }
 
     @Override
     public List<Personal> allPersonal() {
         logger.debug("called allPersonal() method. UserServiceImpl.class");
-        return daoFind.selectAllPersonal();
+        return daoPersonal.selectAllPersonal();
     }
 
 
@@ -39,45 +31,45 @@ public class UserServicePersonalImpl implements UserServicePersonal {
     public void createPersonal(String firstName, String lastName, int bossId, int premium, int salary, int jobId,
                                int department_id, Integer patient_id, String username, String password) {
         logger.debug("called createPersonal() method. UserServiceImpl.class");
-        daoChange.createPersonal(firstName, lastName, bossId, premium, salary, jobId,
+        daoPersonal.createPersonal(firstName, lastName, bossId, premium, salary, jobId,
                 department_id, patient_id);
-        daoChange.createPersonalData(username, password, jobId);
+        daoPersonal.createPersonalData(username, password, jobId);
     }
 
     @Override
     public void updatePersonal(int id, String lastName, int bossId, int premium, int salary, int jobId, int department_id, Integer patient_id) {
         logger.debug("called updatePersonal() method. UserServiceImpl.class");
-        daoChange.updatePersonal(id, lastName, bossId, premium, salary, jobId, department_id, patient_id);
+        daoPersonal.updatePersonal(id, lastName, bossId, premium, salary, jobId, department_id, patient_id);
     }
 
 
     @Override
     public void deletePersonal(int id) {
         logger.debug("called deletePersonal() method. UserServiceImpl.class");
-        daoChange.deletePersonal(id);
+        daoPersonal.deletePersonal(id);
     }
 
     @Override
     public List<Personal> getByIdPersonalList(int id) {
         logger.debug("called getByIdPersonalList() method. UserServiceImpl.class");
-        return daoFind.findByIdList(id);
+        return daoPersonal.findByIdList(id);
     }
 
     @Override
     public Personal getByIdPersonal(int id) {
         logger.debug("called getByIdPersonal() method. UserServiceImpl.class");
-        return daoFind.findByIdPersonal(id);
+        return daoPersonal.findByIdPersonal(id);
     }
 
     @Override
     public Personal getByUserNamePersonal(String name) {
         logger.debug("called getByUserNamePersonal() method. UserServiceImpl.class");
-        return daoFind.getByName(name);
+        return daoPersonal.getByName(name);
     }
 
     @Override
     public List<Personal> findByLastName(String lastName) {
         logger.debug("called findByLastName() method. UserServiceImpl.class");
-        return daoFind.findByLastName(lastName);
+        return daoPersonal.findByLastName(lastName);
     }
 }
