+++ src/main/resources/applicationContext.xml
@@ -6,5 +6,5 @@
         http://www.springframework.org/schema/beans/spring-beans.xsd
         http://www.springframework.org/schema/context
         http://www.springframework.org/schema/context/spring-context.xsd">
-    <context:component-scan base-package="com.test.dao" />
+    <context:component-scan base-package="com.test.dao.daoImpl" />
 </beans>
